+++ src/ai_image_renamer/__init__.py
@@ -4,7 +4,7 @@
 #  https://www.kolja-nolte.com
 #  kolja.nolte@gmail.com
 #
-#  This work is licensed under the MIT License. You are free to use, modify, and distribute this work, provided that you include the copyright notice and this permission notice in all copies or substantial portions of the work. For more information, visit: https://opensource.org/licenses/MIT
+#  This work is licensed under the MIT License. You are free to use, modify, and distribute this work, provided that you include the copyright notice and this permission notice in all copies or substantial portions of the work. For more information, visit: https://opensource.org/licenses/MIT  # noqa: E501
 #
 #  @author      Kolja Nolte
 #  @email       kolja.nolte@gmail.com
@@ -14,12 +14,13 @@
 #  @repository  https://gitlab.com/thaikolja/ai-image-renamer
 
 """Package init and public API exports."""
+# ruff: noqa: E402
 
 from importlib.metadata import PackageNotFoundError, version
 
 # Package metadata constants
-_PACKAGE_NAME = 'ai-image-renamer'
-_FALLBACK_VERSION = '1.1.0'
+_PACKAGE_NAME = "ai-image-renamer"
+_FALLBACK_VERSION = "1.3.0"
 
 # Try to fetch version from installed package metadata
 try:
@@ -30,15 +31,15 @@ except PackageNotFoundError:
 # Import public API components
 from .cli import main
 from .renamer import ImageRenamer
-from .utils import verify_image_file, encode_image, sanitize_image_path, get_words
+from .utils import encode_image, get_words, sanitize_image_path, verify_image_file
 
 # Define public interface for star imports
 __all__ = [
-    '__version__',
-    'ImageRenamer',
-    'main',
-    'verify_image_file',
-    'encode_image',
-    'sanitize_image_path',
-    'get_words',
+    "__version__",
+    "ImageRenamer",
+    "main",
+    "verify_image_file",
+    "encode_image",
+    "sanitize_image_path",
+    "get_words",
 ]